requirements.txt
Django==4.2.16
pymysql
djangorestframework
djangorestframework-simplejwt
dj-rest-auth 
django-allauth 
social-auth-app-django
django-filter
openpyxl
psycopg2-binary
django-cors-headers
Pillow
kafka-python
httpx